Add translation support
I tried to do it with jinja2's i18n and Babel stuff, but it was just too
complicated, so I made my own system. Maybe one day I'll figure out how
to make that other stuff work, but it is not this day. It works, so I
don't care.